requirements.txt
Django==2.0
django-crispy-forms==1.7.0
django-markdown-deux==1.0.5
django-pagedown==0.1.3
djangorestframework==3.7.7
Markdown==2.6.10
markdown2==2.3.5
olefile==0.44
Pillow==4.3.0
pytz==2017.3
